fix mask

+++ resources/utils/image_utils.py
@@ -183,7 +183,7 @@ def convert_mask_to_png(filename, raster, metadata, colours, idx):
     new_raster = new_raster.astype('uint8')
 
     with rio.open(UPLOAD_DIRECTORY + png_filename, 'w', **new_metadata) as dst:
-        dst.write(new_raster)
+        dst.write(new_raster.astype('uint8'))
 
     return png_filename
